🔨 라우터 분리 및 query,styles 관련 파일 App.tsx로 이동
