+++ src/App.tsx
@@ -1,62 +1,19 @@
-import { PATH } from '@/constants/path'
-import { createBrowserRouter, Navigate, Outlet, RouterProvider } from 'react-router-dom'
+import router from '@/routes'
+import { RouterProvider } from 'react-router-dom'
+import GlobalStyles from '@/styles/GlobalStyles'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 
-import CreatePlaylistPage from '@/pages/CreatePlaylistPage'
-import FollowPage from '@/pages/FollowPage'
-import LoginPage from '@/pages/LoginPage'
-import Page404 from '@/pages/Page404'
-import PlaylistPage from '@/pages/PlaylistPage'
-import ProfilePage from '@/pages/ProfilePage'
-import SearchPage from '@/pages/SearchPage'
-import RootLayout from '@/layout/Root'
-import ChatPage from '@/pages/ChatPage'
-import HomePage from '@/pages/HomePage'
-import ResetPwPage from '@/pages/ResetPwPage'
-import SignUpPage from '@/pages/SignUpPage'
-import checkAuth from '@/service/auth/checkAuth'
+const queryClient = new QueryClient()
 
-const PrivateRoute = () => {
-  return checkAuth() ? <Outlet /> : <Navigate to={PATH.LOGIN} replace />
+const App = () => {
+  return (
+    <QueryClientProvider client={queryClient}>
+      <GlobalStyles />
+      <RouterProvider router={router} />
+      <ReactQueryDevtools initialIsOpen={false} />
+    </QueryClientProvider>
+  )
 }
-
-const router = createBrowserRouter([
-  {
-    path: PATH.HOME,
-    element: <RootLayout />,
-    errorElement: <Page404 />,
-    children: [
-      {
-        path: PATH.LOGIN,
-        element: <LoginPage />,
-      },
-      {
-        path: PATH.SIGNUP,
-        element: <SignUpPage />,
-      },
-
-      { path: PATH.EDITPW, element: <ResetPwPage /> },
-      {
-        element: <PrivateRoute />,
-        children: [
-          {
-            index: true,
-            element: <HomePage />,
-          },
-          { path: PATH.PLAYLIST, element: <PlaylistPage /> },
-          { path: PATH.USER_PROFILE, element: <ProfilePage /> },
-          { path: PATH.CREATEPLAYLIST, element: <CreatePlaylistPage /> },
-          { path: PATH.FOLLOW, element: <FollowPage /> },
-          { path: PATH.CHAT, element: <ChatPage /> },
-          {
-            path: PATH.SEARCH,
-            element: <SearchPage />,
-          },
-        ],
-      },
-    ],
-  },
-])
-
-const App = () => <RouterProvider router={router} />
 
 export default App
